pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "premiere-suites-scraper"
version = "2.0.0"
description = "A comprehensive web scraper for extracting property data from Premiere Suites"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["web-scraping", "real-estate", "vector-database", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP :: Browsers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.2",
    "selenium>=4.15.2",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "scrapy>=2.11.0",
    "lxml>=4.9.3",
    "html5lib>=1.1",
    "fake-useragent>=1.4.0",
    "python-dotenv>=1.0.0",
    "chromedriver-autoinstaller>=0.6.2",
    "webdriver-manager>=4.0.1",
    "langchain>=0.0.350",
    "langchain-community>=0.0.1",
    "sentence-transformers>=2.2.2",
    "faiss-cpu>=1.7.4",
    "chromadb>=0.4.18",
    "openai>=1.3.7",
    "tiktoken>=0.5.1",
    "pydantic>=2.5.0",
    "tqdm>=4.66.1",
    "reportlab>=4.0.7",
    "fpdf2>=2.7.6",
    "Pillow>=10.1.0",
    "markdown>=3.5.1",
    "qdrant-client>=1.7.0",
    "flask>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
premiere-scraper = "main:main"

[project.urls]
Homepage = "https://github.com/yourusername/premiere-suites-scraper"
Repository = "https://github.com/yourusername/premiere-suites-scraper"
Issues = "https://github.com/yourusername/premiere-suites-scraper/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.txt", "*.md", "*.json"]

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 100
extend-ignore = ["E501", "W503"]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "build",
    "dist",
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
